TRQ-3239 - track signal sending to pids

+++ src/resmom/linux/mom_mach.c
@@ -2627,6 +2627,7 @@ int mom_set_use(
 
 int kill_task(
 
+  job  *pjob,   /* I */
   task *ptask,  /* I */
   int   sig,    /* I */
   int   pg)     /* I (1=signal process group, 0=signal master process only) */
@@ -2793,10 +2794,37 @@ int kill_task(
             log_record(PBSEVENT_JOB, PBS_EVENTCLASS_JOB, ptask->ti_qs.ti_parentjobid, log_buffer);
             
             if (pg == 0)
-              kill(ps->pid, SIGTERM);
+              {
+              /* make sure we only send a SIGTERM one time per process */
+              if ((ps->state != 'Z') &&
+                  (pjob->ji_sigtermed_processes->find(pid) == pjob->ji_sigtermed_processes->end()))
+                {
+                sprintf(log_buffer, "%s: killing pid %d task %d gracefully with sig %d",
+                  __func__, ps->pid, ptask->ti_qs.ti_task, SIGTERM);
+
+                log_record(PBSEVENT_JOB, PBS_EVENTCLASS_JOB,
+                    ptask->ti_qs.ti_parentjobid, log_buffer);
+
+                kill(ps->pid, SIGTERM);
+                pjob->ji_sigtermed_processes->insert(ps->pid);
+                }
+              }
             else
-              killpg(ps->pid, SIGTERM);
-            
+              {
+              if ((ps->state != 'Z') &&
+                  (pjob->ji_sigtermed_processes->find(pid) == pjob->ji_sigtermed_processes->end()))
+                {
+                sprintf(log_buffer, "%s: killing pid %d task %d gracefully with sig %d",
+                  __func__, ps->pid, ptask->ti_qs.ti_task, SIGTERM);
+
+                log_record(PBSEVENT_JOB, PBS_EVENTCLASS_JOB,
+                    ptask->ti_qs.ti_parentjobid, log_buffer);
+
+                killpg(ps->pid, SIGTERM);
+                pjob->ji_sigtermed_processes->insert(ps->pid);
+                }
+              }
+ 
             for (i = 0;i < 20;i++)
               {
               /* check if process is gone */
@@ -2856,11 +2884,32 @@ int kill_task(
                   __func__, ps->pid, ptask->ti_qs.ti_task, sig);
                 
                 log_record(PBSEVENT_JOB, PBS_EVENTCLASS_JOB, ptask->ti_qs.ti_parentjobid, log_buffer);
-                
+
                 if (pg == 0)
-                  kill(ps->pid, sig);
+                  {
+                  if (sig != SIGTERM)
+                    {
+                    kill(ps->pid, sig);
+                    }
+                  /* make sure we only send a SIGTERM one time */
+                  else if (pjob->ji_sigtermed_processes->find(ps->pid) == pjob->ji_sigtermed_processes->end())
+                    {
+                    killpg(ps->pid, SIGTERM);
+                    pjob->ji_sigtermed_processes->insert(ps->pid);
+                    }
+                  }
                 else
-                  killpg(ps->pid, sig);
+                  {
+                  if (sig != SIGTERM)
+                    {
+                    killpg(ps->pid, sig);
+                    }
+                  else if (pjob->ji_sigtermed_processes->find(ps->pid) == pjob->ji_sigtermed_processes->end())
+                    {
+                    killpg(ps->pid, SIGTERM);
+                    pjob->ji_sigtermed_processes->insert(ps->pid);
+                    }
+                  }
                 }
               }    /* END if ((ps = get_proc_stat(ps->pid)) != NULL) */
             }      /* END if (i >= 20) */
